@@ -1,1 +1,2 @@
-String baseUrl = 'https://bwabank.my.id/api';
+// String baseUrl = 'https://bwabank.my.id/api';
+String baseUrl = 'https://62795e393620.ngrok-free.app/api';
