@@ -6,11 +6,11 @@ class HomeServiceItem extends StatelessWidget {
   final String title;
   final VoidCallback? onTap;
   const HomeServiceItem({
-    Key? key,
+    super.key,
     required this.iconUrl,
     required this.title,
     this.onTap,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
